visionmodel/requirements.txt
pandas
numpy
matplotlib
mplfinance
tqdm
scikit-learn
torch
torchvision
Pillow
yfinance